Avoid errors when requests exclude the response param
This is one way of addressing #1654, although @floehopper discusses an
alternative approach in 28785dd848d206e80e82c6c147b0db719258862a.

The problem we were seeing was caused by a combination of:

* The `parse: Integer` option being set on a `value_question`.

* The presence of a `nil` response in the `SmartAnswerPresenter#all_responses`
array

This fix feels OK to me as the current implementation implies that the presence
of `:next` in the params should mean that `:response` is also present.

The effect of this change is for the
`SmartAnswersController#redirect_response_to_canonical_url` behaviour to kick
in and redirect the user back to the URL without the `next` parameter.