Allow direct instantiation of flows in integration tests
Now that most smart answer flows are defined in a class, there doesn't seem to
be much justification for using `SmartAnswer::FlowRegistry#find` to lookup
flows in the integration tests. It seems simpler to have each integration test
require the flow class they need and have the helper instantiate the flow.