Downgrade google-api-client gem to fix analytics rake task
The `analytics:analyse` rake task makes use of the `google-api` command-line
tool which is only available in pre-v0.7 versions of the `google-api-client`
gem. I

I also had to explicitly restrict the version of the `jwt` gem to ~> 0.1.4 to
get the google-api CLI tool to run.

I still haven't got the rake task to work completely, because I'm waiting on
some permission changes to my Google account. However, I'm confident that we
need these changes if the task is ever going to work and they don't seem as if
they should affect any production code.