Fix TypeError for response parsed by Integer() method
We were seeing the following exception [1] in production for value question
responses parsed by `Integer()`:

    TypeError: can't convert nil into Integer

I had mentioned that this might become a problem in this commit [2].

@chrisroos attempted an alternative fix in #1654, but ran into some unexpected
test failures relating to checkbox questions.

Even though I don't think we've seen the relevant exception in production,
I've implemented the same solution for value question responses parsed by
`Float()` to keep the code as consistent as possible.

I think it would be more consistent to rescue `ArgumentError` in
`SmartAnswer::Question::Value#parse_input` and raise a
`SmartAnswer::InvalidResponse`, so that `SmartAnswer::Flow#process` doesn't
have to rescue `ArgumentError`, but that's a separate issue.

Fixes #1654.

[1]: https://errbit.production.alphagov.co.uk/apps/533c35ae0da1159384044f5f/problems/555dc6746578635c07b60300
[2]: 28785dd848d206e80e82c6c147b0db719258862a