Add support for ERB templates for Outcome titles
I've introduced some duplication between the rendering of the Outcome title and
body. I plan to DRY it all up shortly.

In adding support for these ERB templates I realised that
`OutcomePresenter#body` shouldn't be raising an exception when the template
doesn't exist. The existing YAML templates don't error and don't render
anything when either the title or body are missing.