Check for missing outcome artefacts in regression tests
It's possible that a change in the flow and its corresponding test data may
mean that no outcome artefact is generated for one that exists in the git repo.
However, as the code stood, we would not detect this, because the assertion
diffs were only being done for the individual outcome artefacts. This change
means that we do a diff for the directory as a whole after all the response
combinations have been tried and the assertion fails if there are any
differences.

Unfortunately, because Minitest doesn't have an equivalent of RSpec's
`after(:all)`, we have to use some slightly nasty code to run our new test
at the end of all the tests for a given flow.