Hack-fix failing autoloading of flows.
Now that most of the smart answer flows are classes
as of 0800f52b001f3d0d4bed9a497b7ab0aa2ce60f9d, in development environment
we often get uninitialized constant SmartAnswer::<FlowName> error.

This is because the flows dir structure and namespaces do not follow
rails conventions and autoloading fails. This is a very hackish solution
that reloads the class every time it is being loaded in dev environment.