Extend regression test coverage for state-pension-topup flow
These response combinations give full test coverage of the flow & calculator
code, except for:

* Validation for `money_question :how_much_extra_per_week?`. However, it not
currently possible to test error states and in any case doing so is not
relevant to introducing ERB outcome templates which is the current aim of
introducing these regression tests.

* The `total = 0` line in the `else` condition in
`StatePensionTopupCalculator#lump_sum_amount`. However, this code appears to be
unreachable via the user interface.

Test coverage alone wasn't sufficient to generate all the possible outcomes, so
I had to generate a suitable range of responses by inspecting the code.

@chrisroos plans to introduce a check that all outcomes have been generated, so
if I've missed any that should be caught then.