Update responses for benefit-cap-calculator
I've arbitrarily chosen 50 as the response for each of the "How much..."
questions.

I've combined all of the responses to the checkbox_question about received
additional benefits. The array of chosen benefits is used to decide whether to
display the corresponding question about the amount received for that benefit,
so there's no advantage to listing them all separately.

I've added a single benefit (bereavement) as a response to the
checkbox_question so that I can exercise the
`receiving_non_exemption_benefits?` node.

I've added 350 to the `housing_benefit_amount` question to exercise the `else`
branch of the `if new_housing_benefit_amount < 0.5` condition in the
`outcome_affected_greater_than_cap` outcome.