List only the nodes that aren't being exercised
This regression test fails when it detects that our question-and-responses data
doesn't contain enough responses to exercise every node in the flow.

It was previously comparing an array of expected nodes with an array of
exercised nodes. While the test would fail as expected, the output (the
contents of both arrays were displayed) made it hard to see what was going.

This change should make it much easier to see which additional nodes need to be
exercised so that you can add responses as appropriate.