Update responses for calculate-agricultural-holiday-entitlement
I've arbitrarily chosen a holiday start date of 1st Feb 2015 as Time is frozen
to 1st Jan 2015 in both generator script and regression test.

I've chosen a low and high number for both the `how_many_total_days?` and
`how_many_weeks_at_current_employer?` questions, in the hope that this helps me
exercise all of the Smart Answer.

This set of responses gives me 100% coverage of the flow and nearly 100%
coverage of the `AgriculturalHolidayEntitlementCalculator`. The only line
that's not covered is the `if` branch of the `#calculation_period` method as
our test data and regression test currently hardcodes the current date to 1st
Jan 2015. @floehopper pointed out that it doesn't matter in this case because
the outcome of this method only affects calculations in other methods: it
doesn't cause different branches of code to be executed.