Reduce responses for register-a-death test
This reduced set of countries is the minimal set I could come up with which
provides almost full coverage of the register-a-death smart answer flow.

The following bits of logic appear to be unreachable:

* In the `embassy_result` outcome, in the `:documents_required_embassy_result`
calculate block, the special case where the country of death is Libya.

* In the `embassy_result` outcome, in the `:fees_for_consular_services`
calculate block, the special case where the current location is Libya.

* In the `embassy_result` outcome, in the `:footnote` calculate block, the
special case for eastern caribbean countries which generates the
`:footnote_caribbean` phrase key.