Freeze time when generating responses and expected results
This matches the freezing of time in the smart_answers_regression_test.

I noticed a problem where some of the rendered pages for
calculate-agricultural-holiday-entitlement included validation errors. We
should only be rendering outcomes, which can never be in an error state, so
this was unexpected.

The problem was caused by us using the current date when generating the
responses-and-expected-results, and a fixed date of 1st Jan 2015 when running
the regression tests. Responses that lead to valid outcomes based on the
current date don't necessarily lead to valid outcomes when running on the 1st
Jan 2015. Freezing time in this way should avoid problems like this in future.

The next step will be to regenerate the responses-and-expected outcomes for all
of the existing regression tests. If those files change then I'll also need to
regenerated the rendered outcomes and checksum data.