Update calculate-married-couples-allowance-questions-and-responses.yml
* I've chosen different dates for `:whats_the_husbands_date_of_birth?` and
`:whats_the_highest_earners_date_of_birth?` based on the assumption that
this'll cause different branches to be exercised.

* I've chosen different amounts (50,000 and 55,000) for
`:whats_the_husbands_income?` and `:whats_the_highest_earners_income?` based on
the assumption that this'll cause different branches to be exercised.

* I've added 20,000 as response to `:whats_the_husbands_income?` so that I
reach the `:husband_done` node.

* I've added 20,000 as response to `:whats_the_highest_earners_income?` so that
I reach the `:highest_earner_done` node.

* I've added 27,001 as a response to `:whats_the_husbands_income?` so that I
reach the `else` branch of the `if remaining_reduction > 0` condition in
`MarriedCouplesAllowanceCalculator#calculate_allowance`. It took me quite a lot
of trial and error to get to this branch of the code and I'm not sure I can do
a good job of explaining it. I'm fairly confident that it's because the income
(27001) is greater than the `income_limit_for_personal_allowances` figure for
2014/15, but can't explain what the upper threshold would be that would cause
`remaining_reduction` to be greater than 0 and therefore for the `if` branch to
be reached.

* I've chosen arbitrary amounts for the remaining three questions.

These responses give me close to 100% coverage of both the
calculate-married-couples-allowance flow and the
married_couples_allowance_calculate. The only two lines not exercised are
validation errors, and we're not currently interested in those in our
regression tests.