Move married_couples_allowance_calculator.rb
This calculator was added on 13th Apr 2012[1] which was before the
lib/smart_answer/calculators directory was added on 14th Jun 2012[2]. I can't
think of a good reason to keep it outside of the calculators directory.

I've moved the corresponding unit test and moved both the calculator and unit
test into the `SmartAnswer::Calculators` namespace.

[1]: 69af6847ba08f41756fd02b8a0ee4e753d90f032
[2]: 8ede18b30e8a5983b0a73a35dba7f1680b0b503c