Allow validating date question answer if input is in allowed range
The initial idea was to enforce this validation for all date
questions. However, we have a lot of questions that have a moving
window of dates shown in the dropdowns ie
from { 1.year.ago }

In theory this would allow users to save the link for later. If such
a link would be clicked later, it would show a validation warning,
which is wrong, because in most of the cases we can still show the
correct answer.

Therefore the validation is optional. It uses from and to dates to
enforce the range. It also takes into account cases where
from > to (this is done to change the ordering of years in the date
selector).