Update calculate-state-pension-questions-and-responses.yml
I started out with a fairly arbitrary set of amounts and then revised them as I
learnt from the code coverage reports.

The statements below assume that time is frozen to 1st Jan 2015, which we're
currently doing in the generate-responses-and-expected-results script and the
smart_answers_regression_test.

* Add 1980-01-01 as date of birth response. Given that we freeze time to 1st
Jan 2015, this puts the user at 35 years old, i.e. before state pension age.
This exercises the `:years_paid_ni?` node.

* Add 2000-01-01 as date of birth response. Given that we freeze time to 1st
Jan 2015, this puts the user at 15 years old which allows us to exercise the
`:too_young` node.

* Add 1950-02-01 as date of birth response. This puts the state pension date at
1st Feb 2015. Given that we freeze time to 1st Jan 2015, this means that we're
before the state pension date and the date is within four months of the state
pension date. All of which means that we reach the `:near_state_pension_age`
node.

* Add 1954-01-01 as a response to `:dob_amount?` node. This satisfies the
`:woman_and_born_in_date_range?` predicate which causes us to exercise the
`:pay_reduced_ni_rate?` node.

* Add 1951-04-06 as a response for `dob_age?`. Given that we freeze time at 1st
Jan 2015, this means that `state_pension_date` for males is 6th Apr 2016 (later
than today and equal to 6th Apr 2016) but that the `pension_credit_date` is 6th
May 2012 (earlier than today). This exercise line 94 of calculate-state-pension
as of this commit.

* Add 1960-01-01 as a response to `:dob_amount?` to exercise the code in the
`:automatic_ni?` predicate in the `:received_child_benefit?` node.

* Add 11 a response to `:years_paid_ni?` to exercise the code in the
`:new_rules_and_less_than_10_ni?` predicate in the `:received_child_benefit?`
node.

* Add 1951-01-01 as dob and 30 as NI years so that we go straight from
`:years_paid_ni?` to `:amount_result`. This ensures we exercise line 533
(`qualifying_years + calc.calc_qualifying_years_credit(0)`) of the
calculate-state-pension flow.

* Add 1951-06-01 as date of birth, which means that `auto_years_entitlement` is
true and we exercise the code in line 625.

* Adding 23 means that `enough_qualifying_years` is false and
`enough_remaining_years` is true which means we exercise the code on line 629.

* Adding 1950-02-01 so that `within_four_months_one_day_from_state_pension?` is
true and we exercise the code in the `if` branch that's on line 590.

* 1952-02-29 is a leap year which means we execute line 85 in
state_pension_amount_calculator

* 1992-04-06 means that `#at_least_two_year_credit_age?` is true

* 1993-04-06 means that `#at_least_one_year_credit_age?` is true

* Add 0 and 1 as years worked to execute lines 142 to 146 of
`#calc_qualifying_years_credit` in state_pension_amount_calculator.

NOTE. I'm fairly confident that line 100 of calculate-state-pension (at the
time of this commit) can't be reached. I don't think it's possible to have a
`state_pension_date` that's earlier than 6th Apr 2016 _and_ a
`pension_credit_date` that's later than today.

NOTE. I don't think line 494 (the `if` branch in the `:near_state_pension_age`
outcome) is reachable, for the same reasons that line 100 isn't reachable.

NOTE. I'm fairly confident that `if Date.today < calc.state_pension_date - 35`
on line 597 can never be true because
`calc.within_four_months_one_day_from_state_pension?` has to be true to get to
that line, and that means that `calc.state_pension_date` is within the next
four months.

NOTE. `StatePensionAmountCalculator#enough_qualifying_years?` isn't covered but
that's because it's not actually called from anywhere.