Validate user input is in allowed date range for some answers
I picked some questions that can fail if the date is out of range.
e.g. the state pension calculator was erroring when a user tweaked
the url to add a DOB that is not covered by our logic.

Now if a user tries to modify the date that is out of range provided
by the UI (e.g. tweaking URL or using web inspector to change values
in form) it will bring the user back to the date question with invalid
value asking to provide a valid answer.