Make COUNTRIES_WITH_CONSULATE_GENERALS an array c.f. other consts
I found it confusing that this constant was a space-separated set of strings
combined into a single string whereas all the other similar constants are
arrays of strings. I assume it was just a typo.

This commit shouldn't change any behaviour, because all the uses of the
constant go via the
`SmartAnswer::Calculators::RegistrationsDataQuery#has_consulate_general?`
method.