Optional coverage report when generating expected results
This is useful when trying to select a subset of responses which still give
decent coverage of the logic in a smart answer flow.

I've chosen the default formatter, rather than the `rcov` one we use in the
main test suite, because the default one gives a bit more information and we
don't need Jenkins compatibility, because this script is not run as part of
the CI build.