Fix a bug with answer changing in smartdown.
Problem as reported:
--------------------

Three issues with the Parental Leave & Pay calculator: two minor, one serious, all very likely related.

From the results page at:

https://www.gov.uk/pay-leave-for-parents/y/no/2015-6-8/employee/yes/yes/160-week/yes

* clicking [change] for mother's employment details.

has no effect - should return us to the correct form.

Similarly,

https://www.gov.uk/pay-leave-for-parents/y/yes/2015-5-5/employee/employee/yes/yes/200-week/yes/yes/yes/150-week/yes

* clicking [change] for partner's employment details

has no effect - as above

But

* clicking [change] for mother's employment details

does take us back to the correct form. If we then

* change the mother's earnings
* click [next_step]

we get back to the results page, but it's the *partner's* earnings that are updated,
not the mother's.

Bug description:
----------------

When a change link was opened, responses to the question that's being changed
(previous_response_* query params) were added to the current state, which is
incorrect. The current state is being calculated based purely on those answers,
therefore we were displaying the question that goes after the one that is being
changed.