Use Erubis instead of ERB for outcome templates
The ERB trim mode chosen in [1] was intended to replicate how Rails handles
whitespace trimming. However, in fact Rails uses Erubis [2] and its default
trim mode.

This commit uses Erubis in its default trim mode instead of ERB to render
outcome templates.

This change means there is no longer any need to use `-%>` tags on lines which
(other than the ERB tag) only contain whitespace. This is because the
suppression of leading & trailing whitespace and the trailing newline is the
default behaviour in Erubis.

I had to add an explicit newline to one of the outcome templates in order to
generate the appropriate paragraph HTML tag to match the regression test
outcome HTML.

This change will make it easier to switch to Rails templating in the future
which is part of our plan.

[1]: d0520fcce0529f40a103853ecc3f7a09d287976f
[2]: http://api.rubyonrails.org/classes/ActionView/Base.html