Refactor a complex conditional
The else branch that was refactored is only reached for residents in
the UK (third country variant leads to a differnt outcome, ceremony
country is covered by the if of this else). That eliminates one branch
of code completely. Using the predicate logic and the fact above I
was able to simplify the conditional without changing behaviour (I
hope so at least). The outcome does not make much sense, but it seems
to be correct unless FCO tells us so.