Build outcome template directory relative to flow load path
I want to add some integration tests which run against flows which are defined
in `test/fixtures/smart_answer_flows` as opposed to the usual
`lib/smart_answer_flows`.

Rather than introduce a new global setting, we can make use of the existing
`SmartAnswer::FlowRegistry#load_path` method whose value is changed in the
relevant integration tests.

Since the `FlowRegistry` is in the `SmartAnswer` namespace, I elected to access
it from within the `Outcome` class which is itself in that namespace (whereas
`OutcomePresenter` is not).

Since `Outcome#flow_name` was only needed to build the template directory, we
can replace it with `Outcome#template_directory`.

I've made `Outcome#template_directory` return an instance of `Pathname` to make
it easier to join to individual template filenames.