Update calculate-your-holiday-entitlement-questions-and-responses.yml
I added arbitrary values to start with.

I initially chose some dates in 2014 as we freeze time to 2015-01-01 for the
regression tests.

I added some dates in 2012 to exercise the leap year code in
holiday_entitlement.rb (methods relating to `feb29th_`).

I chose 1680 for `annualised_hours` as it's 35 hours multiplied by 48 weeks.

I initially added '1' to `how_many_days_per_week?` but subsequently added 7 to
exercise line 71 (as of this commit) of calculate-your-holiday-entitlement.

I've got 100% coverage of the flow and nearly 100% of the holiday_entitlement
calculator. The only line of holiday_entitlement that's not exercised is the
line calling `super` in the `else` branch within `#method_missing`.