Remove unused code branch
This code branch was present in the CNI outcome. No third country
outcomes are leading to in anymore, so the remaining two scenarions
(resident of the UK or the ceremony country) are always going to be
true. Therefore this clause can be removed leaving only the `if`
block contents and deleting the else block.

Then the nested conditional with cote-d-ivoire was bubbled up to
reduce nesting.