Always show documents_needed_7_days_residency for consular_cp_outcome
None of the countries leading to consular_cp_outcome is in the
ss_21_days_residency_required_countries list. This may or may not be
coincidental, but keeping dead code around is not good.

This was the last use-case for ss_21_days_residency_required_countries?
method, so it's being removed too.