Update childcare-costs-for-tax-credits-questions-and-responses.yml
I've filled in responses for the money questions. I've answered them all based
on £10/week but will adjust as necessary depending on the code coverage.

I had to add some £0 responses as we have branches of code that explicitly
check for £0.

I had to add £20 as a response to `:new_weekly_costs?` so that the weekly
difference is more than or equal to £10.

I have one line of the childcard-costs-for-tax-credits flow that's not covered.
The `else` branch of `if cost_change_4_weeks` won't ever be reached as
`cost_change_4_weeks` is hardcoded to `true` in the `:old_weekly_amount_2?`
question.

I have two methods in the childcare_cost_calculator that aren't covered,
although it looks like neither of them are used in anything apart from the
tests: `cost_change_annual` and `cost_change_month`.