Remove unreachable code in overseas-passports flow
There are only two questions from which it's theoretically possible to reach
the `:result` outcome:

* `:child_or_adult_passport?`
* `:country_of_birth?`

In both these questions there are a bunch of `next_node_if` statements and a
`next_node` statement inside an `on_condition` block, all followed by a final
`next_node` statement which points at the `:result` outcome.

In both cases the final `next_node` statement within the `on_condition` block
points at the `:ips_application_result` outcome i.e. not the `:result` outcome.

Thus the `:result` outcome is only reachable if the
`SmartAnswer::Calculators::PassportAndEmbassyDataQuery#ips_application?`
predicate returns `false`.

However since "st-helena-ascension-and-tristan-da-cunha" was switched to `type:
ips_application_3` in [1] the predicate only returns `false` when the response
to the `:which_country_are_you_in?` question is one of the following:

* "syria"
  * However, in this case the `PassportAndEmbassyDataQuery#ineligible_country?`
  predicate returns `true` for a response of "syria" and the user is taken
  immediately to the `:cannot_apply` outcome.
* "the-occupied-palestinian-territories"
  * However, in this case the user is subsequently asked to choose either
  "gaza" or "jerusalem-or-westbank" and either of these responses cause the
  `PassportAndEmbassyDataQuery#ips_application?` to return `true`.
* "united-kingdom"
  * However, in this case this response is not included in the
  `:which_country_are_you_in?` question's country select list (it is not
  included in such questions by default; only if the `:include_uk` option is
  set to `true`). So it wouldn't be possible for a user to give this response
  in production without hacking the URL.

Thus I'm pretty convinced the `:result` outcome is unreachable given the
current logic and data. Obviously if the data changed, it might be possible
for the outcome to become reachable again, but I don't think the overhead of
keeping all this code hanging around is justified.

I've elected to remove the two `next_node(:result)` statements which should
mean we will see a `RuntimeError` exception with the message "Next node
undefined" if none of the other conditions are satisfied by a response from
the user.

I've also tried to remove as much code, content & data which the removal of
the outcome has rendered redundant.

[1]: 3a86b8ecd1c4686885dcbd9466da43f504999df9