Make ActionView ERB template search path flow-specific
Previously the `ActionView::Base` constructor was being passed the root (`/`)
of the file system as the path to search for ERB templates. This was necessary,
because the `Tempfiles` created in the test are being created in `/tmp`.

This is obviously undesirable, since there'd be a remote theoretical
possibility that the presenter could be forced to load and evaluate a template
in a file not related to the flow or even outside the project root altogether.

This commit makes the search path much more specific by setting it to the
template directory specific to the flow. The tests still need to override this
template directory, but the code feels safer.