Standardise birth registrations in Gibraltar
Affected URL /register-a-birth/y/gibraltar

It use to refer to Spain for birth registrations in Gibraltar.
As per original ticket:

the outcome for Gibraltar - https://www.gov.uk/register-a-birth/y/gibraltar - needs to be brought into line with the other 'Outcome Commonwealth' outcomes (like Australia, etc) This means that the 2 references to Spain just have to change to 'Gibraltar'.