Update estimate-self-assessment-penalties-questions-and-responses.yml
Added dates that should cover the branches I can see in the flow.

Chose an arbitrary amount for the estimate tax charge.

Added an additional amount of £6300 to exercise a couple of specific branches in the calculator.

The flow is 100% covered using these responses.

The calculator is 89.23% covered:

* Line 23 (`result = 1000`) can't be reached. The maximum value of
`overdue_filing_days` is 181. If we take 92 from that, multiply by 10 and add
100 we get 990, which is less than 1000.

* The `#penalty_interest` method is never used.

* The `#penalty1date` method is never used.

* The `#penalty2date` method is never used.

* The `#penalty3date` method is never used.