Add Democratic Republic of Congo to the list of CNI countries
Affected URL example:
/marriage-abroad/y/democratic-republic-of-congo/third_country/partner_local/opposite_sex

As per original request:

Democratic Republic of Congo is a CNI country. For all opposite marriage
outcomes (for Democratic Republic of Congo) should mirror the advice we
have for Albania.

In relation to opposite marriage outcomes where the customer is a non-UK
resident. They can book an appointment to post notice using the following link:
https://www.consular-appointments.service.gov.uk/fco/#!/british-embassy-kinshasa/notice-of-marriage-or-civil-partnership/slot_picker