Error if we don't have logic to handle the country in marriage-abroad
I've noticed that e.g. removing one country from the CNI list and not
adding it to other lists leads to the outcome_cp_all_other_countries
outcome. This is a same sex civil partnership outcome. It goes there
even if you select opposite sex partner. This is because it was meant
to be a default outcome for SS marriages. Therefore I've scoped all SS
outcomes with a condition block.

That left the opposite sex outcomes with no good default. I don't know
what a good default would be, so it's better to raise a NextNodeUndefined
error and let the developer know something is wrong.