Let developers know if outcome_os_other_countries is used incorrectly
I've noticed that outcome_os_other_countries does not have a default
phraselist and is handling each country separately. Currently it is
possible to add a country to a list that leads to this outcome without
adding handling for the new country. This would lead to an empty outcome.

Raising an error should make it clear what to do in such cases.