Update help-if-you-are-arrested-abroad-questions-and-responses.yml
I added andorra as it's a country that doesn't have any additional download
links in prisoner_packs.yml. This covers line 111 in
help-if-you-are-arrested-abroad.rb.

I added cyprus to satisfy the condition on line 70 of
help-if-you-are-arrested-abroad.rb.

I added iran to exercise the `:answer_two_iran` outcome.

I added syria to exercise the `:answer_three_syria` outcome.

I added albania, austria, greece and spain to cover the full complement of
additional download links.

This set of responses give me 100% coverage of both the
help-if-you-are-arrested-abroad Smart Answer and the arrested_abroad calculator.