Remove unused code from inherits-someone-dies-without-will
Neither outcome_21 nor outcome_22 have been reachable since
9852ac6404e7764169e22e937c214e8b297919dd.

That commit updated the flow so that answering 'england-and-wales' for region
and 'yes' for partner meant that you were automatically taken to an outcome.
This means that it's no longer possible to reach the `parents?` or `siblings?`
nodes having responded to region with 'england-and-wales', and to partner with
'yes'.