Remove unused code from siblings? node
This block of code hasn't been used since
0ce71eeb6bd8da3827ca791f4fd592a28248af1f.

That commit updated the `parents?` question so that responding to region with
'northern-ireland' and to partner with 'yes' took you to a new
`siblings_including_mixed_parents?` question. As of that commit it was no
longer possible to reach the `siblings?` node having responded to region with
'northern-ireland' and to partner with 'yes'.