Remove unreachable code in maternity-paternity-calculator outcome
In running the regression tests it became apparent that `on_payroll` and
`has_contract` will always be set to `yes` if `not_entitled_reason` is not set
given an exhaustive set of responses to the relevant questions.

Thus it's safe to remove these two (untested) lines of code.

Unfortunately the git history is too confusing to work out where,
let alone why, this code exists.