Follow the require_relative convention for fixture_methods
The project seems to use the non-standard tactic of `require_relative` to avoid
having to add the test directory to the load path when running from the command
line e.g. `ruby -Itest test/unit/foo_test.rb`.

While I'm not convinced this is a good idea, it seems wrong to buck the
convention in this one place [1].

[1]: 3431311349d9d78039e7ba3b7b5e7cdc47555a90