Seychelles now offer same sex marriages
The British high commission in the Seychelles is now offering same-sex
consular marriages. The advice should mirror same outcomes we have for
Bolivia same-sex marriage. The Bolivia reference is just a way to convey
the requirements, the two outcomes are not expected to stay in sync over
time.

The worldide organisations API fixture is added via:

rails r script/update-worldwide-location-organisations.rb seychelles