Only run setup once for regression tests
The setup for all these tests should be exactly the same and so we can speed up
the tests a bit by only running the setup block once.

By replacing `require 'webmock/minitest'` with `require 'webmock'` and
explicitly including `WebMock::API` into the test class, we can avoid the
automatic call to `WebMock.reset!` which would normally remove all the stubs
in the test teardown.

By removing the `teardown` method we avoid the call to `Timecop.return` which
means that the one call to `Timecop.freeze` in the setup is enough to stub
time for all the tests.