Introduce partial to DRY up templates
I have to supply the full filename of the partial (including .govspeak.erb) as
we don't have handlers configured that would allow ActionView to find it
automatically.

*NOTE.* I don't have a newline at the end of the _acas_information.govspeak.erb
file so that the rendered outcomes continue to match exactly.

Not supplying the full filename (i.e. just using `render partial:
'acas_information`) results in the following error:

    ActionView::MissingTemplate

    Missing partial /_acas_information with {
      :locale=>[:"en-GB"],
      :formats=>[
        :html, :text, :js, :css, :ics, :csv, :vcf, :png, :jpeg, :gif, :bmp,
        :tiff, :mpeg, :xml, :rss, :atom, :yaml, :multipart_form,
        :url_encoded_form, :json, :pdf, :zip, :gv
      ],
      :variants=>[],
      :handlers=>[:erb, :builder, :raw, :ruby]
    }.
    Searched in:
    * "smart-answers/lib/smart_answer_flows/am-i-getting-minimum-wage"