Convert outcome_affected_greater_than_cap outcome
This outcome has a title and a body.

I moved the title to outcome_affected_greater_than_cap_title.txt.erb.

I introduced the `new_housing_benefit_amount` precalculate block so that I
could use it in the conditional in the new body template, and in the
`new_housing_benefit` precalculate block.

I moved the body to outcome_affected_greater_than_cap_body.govspeak.erb. The
body is an expansion of the following PhraseLists.

* outcome_affected_greater_than_cap_phrase
* housing_benefit_not_zero (depending on `new_housing_benefit_amount`)
* estimate_only
* contact_details