Partially revert f3eea3b8bf1708294afaa40cabda19e8902b385d
@floehopper and @chrisroos paired on this branch.

We're trying to bring the first two commits from #1634 (i.e. the
statutory-sick-pay-updates branch) in to master.

We ran into problems because of conflicts between those two commits and commit
f3eea3b8bf1708294afaa40cabda19e8902b385d that's appeared in master since the
branch was created.

It proved incredibly hard to manually fix the merge conflicts so we're taking a
slightly different approach. We're partially reverting the commit that
introduced the change in master, so that we can cherry pick the two commits
from the branch before re-applying the changes from this reverted commit.

This is how we partially reverted the problematic commit in master:

    $ git co f3eea3b8bf1708294afaa40cabda19e8902b385d~ -- \
      test/integration/smart_answer_flows/calculate_statutory_sick_pay_test.rb

    $ git co f3eea3b8bf1708294afaa40cabda19e8902b385d~ -- \
      lib/smart_answer_flows/calculate-statutory-sick-pay.rb