Remove appeal-a-benefits-decision draft flow
This draft flow hasn't had any logic/content-related changes since Nov 16, 2012
[1] (over 2.5 years ago) when it was moved from `published` to `draft` and
Nov 6, 2012 [2] when it was first (?) published.

Given that there's a significant cost in terms of developer effort in keeping
it around, we've decided to archive it for now. If we find out in the near
future that it is needed, we should be able to bring it back to life without
too much difficulty.

[1]: 915e10a8639065ed278bad21029aef263c8acb1e
[2]: 1cc8903437afa05415d1b4e1ce3b226fc24822ff