Convert tax_credits_cross_border_worker_outcome outcome
The body template includes an expansion of the following PhraseList keys:

* tax_credits_cross_border_going_abroad
* tax_credits_cross_border_already_abroad
* tax_credits_cross_border
* tax_credits_going_abroad_helpline
* tax_credits_already_abroad_helpline

The latter two re-use the partial templates which were introduced previously.

The `:tax_credits_cross_border_worker` precalculate block can be removed now
that the conditional logic is performed by the `<% if is_already_abroad %>`
ERB tag in the outcome template. It's also safe to remove the corresponding
`assert_phrase_list` statements in the integration test, because we have the
regression tests in place.